♻️(frontend) extract base login layout
This allows to reuse the base login page layout for other
authentication related pages (eg. passwor reset)
